portfolio-website/chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-core==0.1.7
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-experimental==0.0.42
langchain-text-splitters==0.0.1
langchain-chroma==0.0.1
langchain-groq==0.0.1
langchain-mistralai==0.0.1
langchain-huggingface==0.0.1
langchain-fireworks==0.0.1
langchain-together==0.0.1
langchain-cohere==0.0.1
langchain-google-genai==0.0.1
python-multipart==0.0.6
requests==2.31.0
cors==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
langgraph==0.0.27
faiss-cpu==1.7.4
PyPDF2==3.0.1
sentence-transformers==2.2.2